+++ src/components/auth/implementation/base.ts
@@ -3,6 +3,7 @@ import * as Reference from "../../reference/implementation.js"
 import * as Storage from "../../storage/implementation.js"
 
 import * as Did from "../../../did/index.js"
+import * as Events from "../../../events.js"
 import * as SessionMod from "../../../session.js"
 import * as Ucan from "../../../ucan/index.js"
 
@@ -108,15 +109,19 @@ export async function register(
 export async function session(
   components: Components,
   authedUsername: Maybe<string>,
-  config: Configuration
+  config: Configuration,
+  eventEmitters: { session: Events.Emitter<Events.Session<Session>> }
 ): Promise<Maybe<Session>> {
   if (authedUsername) {
-    return new Session({
+    const session = new Session({
       crypto: components.crypto,
       storage: components.storage,
+      eventEmitter: eventEmitters.session,
       type: TYPE,
       username: authedUsername
     })
+
+    return session
 
   } else {
     return null