+++ src/fs/filesystem.ts
@@ -128,7 +128,7 @@ export class FileSystem implements API {
         this._publishing = [ cid, true ]
         return this.dependencies.reference.dataRoot.update(cid, proof).then(() => {
           if (this._publishing && this._publishing[ 0 ] === cid) {
-            eventEmitter.emit("publish", { root: cid })
+            eventEmitter.emit("fileSystem:publish", { root: cid })
             this._publishing = false
           }
         })
@@ -772,7 +772,7 @@ export class FileSystem implements API {
     }
 
     this.eventEmitter.emit(
-      "local-change",
+      "fileSystem:local-change",
       { root: await this.root.put(), path }
     )
   }